Refactor juniper stuff
