+++ juniper/src/macros/mod.rs
@@ -1,10 +1,4 @@
-// Wrapper macros which allows built-in macros to be recognized as "crate-local"
-// and helper traits for #[juniper::graphql_subscription] macro.
-
-#[macro_use]
-mod common;
-//#[macro_use]
-//mod interface;
+//! Helper traits for macros.
 
 #[cfg(test)]
 mod tests;
