+++ juniper/src/tests/fixtures/mod.rs
@@ -1,4 +1,4 @@
 //! Library fixtures
 
-///// GraphQL schema and data from Star Wars.
-//pub mod starwars;
+/// GraphQL schema and data from Star Wars.
+pub mod starwars;
