+++ juniper/src/tests/fixtures/starwars/mod.rs
@@ -1,4 +1,4 @@
-pub mod model;
+
 pub mod schema;
 #[cfg(feature = "schema-language")]
 pub mod schema_language;
